+++ CHANGELOG.md
@@ -2,7 +2,13 @@
 
 ### Features
 - `[lf-repository-browser]`: add error handling when `getParentTreeNodeAsync` throws
-- `[lf-repository-browser]`: `setSelectedValuesAsync` clears previously selected items
+- **[BREAKING]** `[lf-repository-browser]` method `setSelectedValuesAsync`: 
+  - change interface to 
+    ```
+    setSelectedNodesAsync: (nodesToSelect: LfTreeNode[], maxFetchIterations: number) => Promise<void>;
+    ```
+  - clears previously selected items that are not included in `nodesToSelect` argument
+  - emits `entrySelected` event for event selected
 
 ### Fixes
 - `[lf-repository-browser]`: fix `entryDblClicked` event firing timing